[REFACT] `ChatService` 리펙토링
