+++ src/main/java/io/oeid/mogakgo/domain/chat/presentation/ChatController.java
@@ -3,7 +3,7 @@ package io.oeid.mogakgo.domain.chat.presentation;
 import io.oeid.mogakgo.common.base.CursorPaginationInfoReq;
 import io.oeid.mogakgo.common.base.CursorPaginationResult;
 import io.oeid.mogakgo.common.swagger.template.ChatSwagger;
-import io.oeid.mogakgo.domain.chat.application.ChatRoomService;
+import io.oeid.mogakgo.domain.chat.application.ChatService;
 import io.oeid.mogakgo.domain.chat.application.dto.res.ChatRoomDataRes;
 import io.oeid.mogakgo.domain.chat.application.dto.res.ChatRoomRes;
 import io.oeid.mogakgo.domain.chat.presentation.dto.res.ChatDataApiRes;
@@ -26,13 +26,13 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/chat")
 public class ChatController implements ChatSwagger {
 
-    private final ChatRoomService chatRoomService;
+    private final ChatService chatService;
 
     @GetMapping
     public ResponseEntity<CursorPaginationResult<ChatRoomRes>> getChatRoomList(
         @AuthenticationPrincipal Long userId,
         @Valid @ModelAttribute CursorPaginationInfoReq pageable) {
-        var result = chatRoomService.findChatRoomsByUserId(userId, pageable.getCursorId(), pageable.getPageSize());
+        var result = chatService.findChatRoomsByUserId(userId, pageable.getCursorId(), pageable.getPageSize());
         return ResponseEntity.ok(CursorPaginationResult.fromDataWithExtraItemForNextCheck(result,
             pageable.getPageSize()));
     }
@@ -42,14 +42,14 @@ public class ChatController implements ChatSwagger {
         @AuthenticationPrincipal Long userId,
         @PathVariable String chatRoomId) {
         return ResponseEntity.ok(
-            chatRoomService.findChatRoomDetailData(UUID.fromString(chatRoomId), userId));
+            chatService.findChatRoomDetailData(UUID.fromString(chatRoomId), userId));
     }
 
     @GetMapping("/{chatRoomId}")
     public ResponseEntity<CursorPaginationResult<ChatDataApiRes>> getChatData(
         @PathVariable String chatRoomId, @AuthenticationPrincipal Long userId,
         @Valid @ModelAttribute CursorPaginationInfoReq pageable) {
-        var result = chatRoomService.findChatMessagesByRoomId(UUID.fromString(chatRoomId), userId,
+        var result = chatService.findChatMessagesByRoomId(UUID.fromString(chatRoomId), userId,
             pageable.getCursorId(), pageable.getPageSize());
         return ResponseEntity.ok(CursorPaginationResult.fromDataWithExtraItemForNextCheck(result,
             pageable.getPageSize()));
@@ -58,14 +58,14 @@ public class ChatController implements ChatSwagger {
     @GetMapping("/project/{projectId}")
     public ResponseEntity<ChatRoomIdApiRes> getChatRoomByProjectId(
         @AuthenticationPrincipal Long userId, @PathVariable Long projectId) {
-        var chatRoomId = chatRoomService.findChatRoomIdByProjectId(projectId);
+        var chatRoomId = chatService.findChatRoomIdByProjectId(projectId);
         return ResponseEntity.ok(ChatRoomIdApiRes.from(chatRoomId.toString()));
     }
 
     @PatchMapping("/{chatRoomId}")
     public ResponseEntity<Void> leaveChatRoom(@AuthenticationPrincipal Long userId,
         @PathVariable String chatRoomId) {
-        chatRoomService.leaveChatRoom(UUID.fromString(chatRoomId), userId);
+        chatService.leaveChatRoom(UUID.fromString(chatRoomId), userId);
         return ResponseEntity.noContent().build();
     }
 }
